extract info components from [name] page
